Calculate the effective sample size

This is the work horse function that calculates the effective sample size. Should usually not be called by the user directly.

effective_samplesize(size, share, corr = 0.5, weights = NULL)

Arguments

size

A vector of sample sizes from different surveys (from different pollsters) for one party.

share

The relative share of votes for party of interest ([0-1])

corr

Assumed correlation between surveys (of different pollsters). Defaults to 0.5.

weights

Additional weights for individual surveys.

Value

A single numeric value: the effective sample size of the pooled sample accounting for the correlation between pollsters.