Replace/prettify matching words/terms in one vector by another
Source: R/prettify.R

The function searches for x values, that occur in current and replaces them with entries in new. Useful for quick renaming/translation of survey column names and by using internal object .trans_df

prettify_strings(x, current = .trans_df$english,
  new = .trans_df$english_pretty)

prettify_de(x)

prettify_en(x)

Arguments

x
A character vector (or factor) that should be renamed.

current
A vector of characters (possibly subset of x). Entries in x that match entries in current will be renamed according to entries in new.

new
A vector of characters that will replace entries in x which have matches in current.

Examples

library(coalitions)
library(dplyr)
# look at sample German federal election polls
surveys <- surveys_sample %>% tidyr::unnest() %>% group_by(pollster) %>% slice(1)
# prettify the polling agency names
prettify_strings(surveys$pollster)

#> [1] "Allensbach"      "Emnid"           "FGW"             "Forsa"          
#> [5] "GMS"             "Infratest dimap" "INSA"           

prettify_en(surveys$pollster)

#> [1] "Allensbach"      "Emnid"           "FGW"             "Forsa"          
#> [5] "GMS"             "Infratest dimap" "INSA"           

prettify_de(surveys$pollster)

#> [1] "Allensbach"        "Emnid"             "Forsch'gr. Wahlen"
#> [4] "Forsa"             "GMS"               "Infratest dimap"  
#> [7] "INSA"