Obtain pooled survey during specified period

Per default, pools surveys starting from current date and going 14 days back. For each pollster within the defined time-frame, only the most recent survey is used.

pool_surveys(surveys, last_date = Sys.Date(), pollsters = c("allensbach",
  "emnid", "forsa", "fgw", "gms", "infratest", "insa"), period = 14,
  corr = 0.5, weights = NULL)

Arguments

surveys
A tibble containing survey results for multiple pollsters as returned by get_surveys.

last_date
Only surveys in the time-window from last_date to last_date - period will be considered for each pollster. Defaults to current date.

pollsters
Character vector of pollsters that should be considered for pooling.

period
See last_date argument.

corr
Assumed correlation between surveys (of different pollsters). Defaults to 0.5.

weights
Additional weights for individual surveys.

Examples

library(coalitions)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

latest <- get_latest(surveys_sample)
pool_surveys(surveys_sample, last_date=as.Date("2017-09-02"))

#> # A tibble: 7 x 8
#>   pollster date       start      end        respondents party  percent votes
#>   <chr>    <date>     <date>     <date>           <dbl> <chr>    <dbl> <dbl>
#> 1 pooled   2017-09-02 2017-08-22 2017-09-02        3055 afd       8.89   272
#> 2 pooled   2017-09-02 2017-08-22 2017-09-02        3055 cdu      38.0   1161
#> 3 pooled   2017-09-02 2017-08-22 2017-09-02        3055 fdp       8.52   260
#> 4 pooled   2017-09-02 2017-08-22 2017-09-02        3055 greens    7.41   226
#> 5 pooled   2017-09-02 2017-08-22 2017-09-02        3055 left      9.06   277
#> 6 pooled   2017-09-02 2017-08-22 2017-09-02        3055 others    4.51   138
#> 7 pooled   2017-09-02 2017-08-22 2017-09-02        3055 spd      23.6    722