Scrape surveys from wahlrecht.de

Scrapes survey tables and performs sanitation to output tidy data

Scrape Bavarian regional polls

Scrape Lower Saxony regional polls

scrape_wahlrecht(address = "https://www.wahlrecht.de/umfragen/emnid.htm",
  parties = c("CDU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW", "AFD",
  "SONSTIGE"))

scrape_by(address = "https://www.wahlrecht.de/umfragen/landtage/bayern.htm",
  parties = c("CSU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW", "AFD",
  "SONSTIGE"))

  scrape_ltw(address = "https://www.wahlrecht.de/umfragen/landtage/niedersachsen.htm",
  parties = c("CDU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW", "AFD",
  "SONSTIGE"))

Arguments

address
http-address from which tables should be scraped.

parties
A character vector containing names of parties to collapse.

Examples

library(coalitions)
library(dplyr)
# select a polling agency from .pollster_df that should be scraped ...
coalitions:::.pollster_df

#> # A tibble: 7 x 2
#>   pollster   address                                             
#>   <chr>      <chr>                                               
#> 1 allensbach https://www.wahlrecht.de/umfragen/allensbach.htm    
#> 2 emnid      https://www.wahlrecht.de/umfragen/emnid.htm         
#> 3 forsa      https://www.wahlrecht.de/umfragen/forsa.htm         
#> 4 fgw        https://www.wahlrecht.de/umfragen/politbarometer.htm
#> 5 gms        https://www.wahlrecht.de/umfragen/gms.htm           
#> 6 infratest  https://www.wahlrecht.de/umfragen/dimap.htm         
#> 7 insa       https://www.wahlrecht.de/umfragen/insa.htm          

# ... here we choose Forsa
address <- coalitions:::.pollster_df %>% filter(pollster == "forsa") %>% pull(address)
scrape_wahlrecht(address = address)

#> Warning:  1 failed to parse.

#> Warning:  3 failed to parse.

#>           date      start        end cdu spd greens  fdp left afd others
#> 1   2018-03-11 2018-03-05 2018-03-09  34  18     13  9.0 10.0  12      4
#> 2   2018-03-05 2018-03-04 2018-03-05  34  19     12  9.0  9.0  13      4
#> 3   2018-03-03 2018-02-26 2018-03-02  35  18     12  9.0  9.0  13      4
#> 4   2018-02-24 2018-02-19 2018-02-23  35  18     12  9.0  9.0  13      4
#> 5   2018-02-17 2018-02-12 2018-02-16  34  16     13  9.0 10.0  13      5
#> 6   2018-02-09 2018-02-07 2018-02-08  31  18     13 10.0 10.0  13      5
#> 7   2018-02-05 2018-01-29 2018-02-02  33  18     13  9.0  9.0  13      5
#> 8   2018-01-29 2018-01-22 2018-01-26  34  18     12  9.0 10.0  12      5
#> 9   2018-01-22 2018-01-22 2018-01-22  34  17     12  8.0 11.0  13      5
#> 10  2018-01-21 2018-01-15 2018-01-19  34  18     12  8.0 10.0  12      6
#> 11  2018-01-18 2018-01-15 2018-01-17  34  18     12  8.0 10.0  12      6
#> 12  2018-01-15 2018-01-08 2018-01-12  33  20     12  9.0  9.0  11      6
#> 13  2018-01-07 2018-01-02 2018-01-05  34  20     12  9.0  9.0  10      6
#> 14  2017-12-27 2017-12-18 2017-12-22  34  19     12  8.0 10.0  12      5
#> 15  2017-12-16 2017-12-11 2017-12-15  33  20     12  8.0 10.0  12      5
#> 16  2017-12-10 2017-12-09 2017-12-10  34  20     12  8.0  9.0  11      6
#> 17  2017-12-10 2017-12-04 2017-12-08  34  20     12  8.0  9.0  11      6
#> 18  2017-12-04 2017-11-27 2017-12-01  34  20     13  8.0  9.0  10      6
#> 19  2017-11-27 2017-11-20 2017-11-24  33  19     12 10.0  9.0  11      6
#> 20  2017-11-20 2017-11-20 2017-11-20  31  21     12 10.0  9.0  12      5
#> 21  2017-11-20 2017-11-13 2017-11-17  32  20     11 12.0  8.0  12      5
#> 22  2017-11-13 2017-11-06 2017-11-10  32  20     10 12.0  9.0  12      5
#> 23  2017-11-06 2017-10-30 2017-11-03  32  21     10 11.0  9.0  12      5
#> 24  2017-11-01 2017-10-23 2017-10-27  33  20     11 11.0  9.0  11      5
#> 25  2017-10-25 2017-10-16 2017-10-20  31  22     10 11.0 10.0  11      5
#> 26  2017-10-18 2017-10-09 2017-10-13  32  20     11 11.0  9.0  12      5
#> 27  2017-10-11 2017-10-02 2017-10-06  32  20     11 11.0  9.0  12      5
#> 28  2017-10-04 2017-09-25 2017-09-29  32  20     10 11.0 10.0  12      5
#> 29  2017-09-22 2017-09-18 2017-09-21  36  22      7  9.5  9.5  11      5
#> 30  2017-09-19 2017-09-11 2017-09-15  36  23      8  9.0 10.0   9      5
#> 31  2017-09-13 2017-09-04 2017-09-08  37  23      8  8.0 10.0   9      5
#> 32  2017-09-08 2017-09-04 2017-09-07  37  23      8  9.0 10.0   9      4
#> 33  2017-09-06 2017-08-28 2017-09-01  38  23      8  8.0  9.0   9      5
#> 34  2017-08-30 2017-08-21 2017-08-25  38  24      7  8.0  9.0   9      5
#> 35  2017-08-23 2017-08-14 2017-08-18  38  24      7  8.0  9.0   9      5
#> 36  2017-08-16 2017-08-07 2017-08-11  39  23      8  8.0  9.0   8      5
#> 37  2017-08-09 2017-07-31 2017-08-04  40  23      8  7.0  8.0   8      6
#> 38  2017-08-02 2017-07-24 2017-07-28  40  22      8  8.0  8.0   8      6
#> 39  2017-07-26 2017-07-17 2017-07-21  40  22      8  8.0  9.0   7      6
#> 40  2017-07-19 2017-07-10 2017-07-14  40  22      8  8.0  9.0   7      6
#> 41  2017-07-11 2017-07-03 2017-07-07  39  22      8  8.0  9.0   8      6
#> 42  2017-07-05 2017-06-26 2017-06-30  39  23      8  8.0  9.0   7      6
#> 43  2017-06-29 2017-06-26 2017-06-28  40  23      9  8.0  9.0   7      4
#> 44  2017-06-28 2017-06-19 2017-06-23  40  23      9  7.0  9.0   7      5
#> 45  2017-06-21 2017-06-12 2017-06-16  39  23      8  8.0 10.0   7      5
#> 46  2017-06-13 2017-06-06 2017-06-09  38  24      8  9.0  8.0   7      6
#> 47  2017-06-07 2017-05-29 2017-06-02  39  24      8  8.0  8.0   7      6
#> 48  2017-05-31 2017-05-22 2017-05-26  38  25      7  9.0  8.0   7      6
#> 49  2017-05-23 2017-05-15 2017-05-19  39  25      7  9.0  8.0   7      5
#> 50  2017-05-17 2017-05-08 2017-05-12  38  26      7  8.0  8.0   7      6
#> 51  2017-05-10 2017-05-02 2017-05-05  36  29      7  7.0  8.0   7      6
#> 52  2017-05-03 2017-04-24 2017-04-28  36  28      8  7.0  8.0   8      5
#> 53  2017-04-26 2017-04-18 2017-04-21  36  30      7  6.0  8.0   9      4
#> 54  2017-04-19 2017-04-10 2017-04-13  36  30      6  6.0  9.0   8      5
#> 55  2017-04-11 2017-04-03 2017-04-07  36  30      7  6.0  8.0   8      5
#> 56  2017-04-04 2017-03-27 2017-03-31  36  29      7  5.0  9.0   8      6
#> 57  2017-03-29 2017-03-20 2017-03-24  34  32      7  6.0  8.0   7      6
#> 58  2017-03-22 2017-03-13 2017-03-17  34  31      7  6.0  7.0   9      6
#> 59  2017-03-15 2017-03-06 2017-03-10  33  32      7  6.0  7.0   9      6
#> 60  2017-03-08 2017-02-27 2017-03-03  33  32      8  6.0  7.0   8      6
#> 61  2017-03-01 2017-02-20 2017-02-24  33  31      8  7.0  7.0   9      5
#> 62  2017-02-22 2017-02-13 2017-02-17  34  31      7  6.0  8.0   8      6
#> 63  2017-02-15 2017-02-05 2017-02-10  34  31      7  5.0  8.0   9      6
#> 64  2017-02-08 2017-01-30 2017-02-03  34  31      8  5.0  8.0  10      4
#> 65  2017-02-01 2017-01-23 2017-01-27  35  26      8  6.0  9.0  11      5
#> 66  2017-01-25 2017-01-16 2017-01-20  37  21     10  6.0  9.0  12      5
#> 67  2017-01-18 2017-01-09 2017-01-13  38  21      9  6.0  9.0  11      6
#> 68  2017-01-11 2017-01-02 2017-01-06  37  20     10  6.0  9.0  12      6
#> 69  2017-01-04 2017-12-27 2017-12-30  37  21     10  6.0  9.0  12      5
#> 70  2016-12-28 2016-12-19 2016-12-23  38  20     10  5.0  9.0  12      6
#> 71  2016-12-20 2016-12-12 2016-12-16  36  22     10  6.0  9.0  11      6
#> 72  2016-12-14 2016-12-05 2016-12-09  37  21     10  6.0  9.0  11      6
#> 73  2016-12-07 2016-11-28 2016-12-02  36  22     11  6.0 10.0  10      5
#> 74  2016-11-30 2016-11-21 2016-11-25  37  22     10  5.0  9.0  11      6
#> 75  2016-11-23 2016-11-14 2016-11-18  36  23     10  5.0 10.0  10      6
#> 76  2016-11-16 2016-11-07 2016-11-11  35  23     11  6.0 10.0  10      5
#> 77  2016-11-09 2016-10-31 2016-11-04  34  23     11  6.0 10.0  11      5
#> 78  2016-11-02 2016-10-24 2016-10-28  34  22     11  7.0 10.0  11      5
#> 79  2016-10-26 2016-10-17 2016-10-21  35  22     11  6.0  9.0  12      5
#> 80  2016-10-19 2016-10-10 2016-10-14  34  22     11  6.0  9.0  12      6
#> 81  2016-10-12 2016-10-04 2016-10-07  34  22     11  6.0 10.0  12      5
#> 82  2016-10-05 2016-09-26 2016-09-30  33  22     11  6.0 10.0  13      5
#> 83  2016-09-28 2016-09-19 2016-09-23  33  23     10  6.0  9.0  14      5
#> 84  2016-09-21 2016-09-12 2016-09-16  33  23     11  6.0  8.0  13      6
#> 85  2016-09-14 2016-09-05 2016-09-09  32  23     11  6.0  8.0  13      7
#> 86  2016-09-07 2016-08-29 2016-09-02  33  22     12  7.0  8.0  12      6
#> 87  2016-08-31 2016-08-22 2016-08-26  33  22     12  6.0  9.0  12      6
#> 88  2016-08-24 2016-08-15 2016-08-19  35  22     12  6.0  9.0  10      6
#> 89  2016-08-17 2016-08-08 2016-08-12  36  21     13  6.0  9.0   9      6
#> 90  2016-08-10 2016-08-01 2016-08-05  36  21     13  5.0 10.0   9      6
#> 91  2016-08-03 2016-07-25 2016-07-29  35  23     12  6.0  9.0   9      6
#> 92  2016-07-26 2016-07-18 2016-07-22  35  22     13  6.0  9.0   9      6
#> 93  2016-07-20 2016-07-11 2016-07-15  35  22     12  7.0  9.0   9      6
#> 94  2016-07-13 2016-07-04 2016-07-08  36  23     12  6.0  9.0   8      6
#> 95  2016-07-06 2016-06-27 2016-07-01  35  21     13  6.0 10.0   9      6
#> 96  2016-06-28 2016-06-20 2016-06-24  34  22     12  6.0  9.0  10      7
#> 97  2016-06-22 2016-06-13 2016-06-17  33  21     13  6.0  9.0  11      7
#> 98  2016-06-15 2016-06-06 2016-06-10  33  21     13  6.0 10.0  11      6
#> 99  2016-06-08 2016-05-30 2016-06-03  34  21     13  7.0 10.0  10      5
#> 100 2016-06-01 2016-05-23 2016-05-27  34  21     13  6.0  9.0  10      7
#> 101 2016-05-24 2016-05-17 2016-05-20  33  20     14  6.0  9.0  12      6
#> 102 2016-05-17 2016-05-09 2016-05-13  34  19     14  7.0  9.0  11      6
#> 103 2016-05-11 2016-05-02 2016-05-06  34  21     13  8.0  9.0  10      5
#> 104 2016-05-03 2016-04-25 2016-04-29  34  21     13  8.0  9.0   9      6
#> 105 2016-04-27 2016-04-18 2016-04-22  33  22     13  8.0  9.0  10      5
#> 106 2016-04-20 2016-04-11 2016-04-15  34  21     13  7.0  9.0  11      5
#> 107 2016-04-13 2016-04-04 2016-04-08  34  21     14  7.0  9.0  10      5
#> 108 2016-04-06 2016-03-29 2016-04-01  35  21     13  7.0  8.0  10      6
#> 109 2016-03-30 2016-03-21 2016-03-24  36  20     13  7.0  8.0  10      6
#> 110 2016-03-22 2016-03-14 2016-03-18  35  20     13  6.0  8.0  13      5
#> 111 2016-03-16 2016-03-07 2016-03-11  35  22     10  7.0 10.0  11      5
#> 112 2016-03-09 2016-02-29 2016-03-04  35  23     11  6.0  9.0  10      6
#> 113 2016-03-02 2016-02-22 2016-02-26  35  24     10  6.0 10.0   9      6
#> 114 2016-02-24 2016-02-15 2016-02-19  35  24     10  6.0 10.0  10      5
#> 115 2016-02-17 2016-02-08 2016-02-12  35  23     11  6.0 10.0   9      6
#> 116 2016-02-10 2016-02-01 2016-02-05  36  24     10  5.0  9.0  10      6
#> 117 2016-02-03 2016-01-25 2016-01-29  36  24     10  5.0  9.0  10      6
#> 118 2016-01-27 2016-01-18 2016-01-22  36  24     10  5.0 10.0  10      5
#> 119 2016-01-19 2016-01-11 2016-01-15  37  23     10  5.0  9.0  10      6
#> 120 2016-01-13 2016-01-04 2016-01-08  38  23     10  5.0 10.0   9      5
#> 121 2016-01-06 2016-12-28 2016-12-30  38  23     10  5.0 10.0   8      6
#> 122 2015-12-29 2015-12-21 2015-12-23  38  23     10  5.0 10.0   8      6
#> 123 2015-12-22 2015-12-14 2015-12-18  39  24     10  4.0 10.0   8      5
#> 124 2015-12-15 2015-12-07 2015-12-11  39  24     11  4.0  9.0   8      5
#> 125 2015-12-09 2015-11-30 2015-12-04  38  24     10  4.0  9.0   8      7
#> 126 2015-12-02 2015-11-23 2015-11-27  39  24     10  5.0  9.0   6      7
#> 127 2015-11-25 2015-11-16 2015-11-22  39  24     10  5.0  9.0   7      6
#> 128 2015-11-17 2015-11-09 2015-11-13  36  25     10  5.0  9.0   8      7
#> 129 2015-11-11 2015-11-02 2015-11-06  38  25     10  5.0  8.0   7      7
#> 130 2015-11-04 2015-10-26 2015-10-30  36  24     11  5.0  9.0   7      8
#> 131 2015-10-28 2015-10-19 2015-10-23  36  24     11  6.0  9.0   6      8
#> 132 2015-10-21 2015-10-12 2015-10-16  38  24     11  4.0 10.0   7      6
#> 133 2015-10-14 2015-10-05 2015-10-09  38  25     10  5.0  9.0   7      6
#> 134 2015-10-07 2015-09-28 2015-10-02  39  25      9  5.0  9.0   7      6
#> 135 2015-09-30 2015-09-21 2015-09-25  40  24     10  5.0 10.0   5      6
#> 136 2015-09-23 2015-09-14 2015-09-18  40  24     10  5.0 10.0   4      7
#> 137 2015-09-16 2015-09-07 2015-09-11  41  24     10  4.0 10.0   4      7
#> 138 2015-09-09 2015-08-31 2015-09-04  41  24     10  4.0 10.0   4      7
#> 139 2015-09-02 2015-08-24 2015-08-28  40  24     11  5.0 11.0   3      6
#> 140 2015-08-26 2015-08-17 2015-08-21  41  23     10  4.0 11.0   4      7
#> 141 2015-08-19 2015-08-10 2015-08-14  43  23     10  4.0 10.0   3      7
#> 142 2015-08-12 2015-08-03 2015-08-07  43  23     10  5.0  9.0   3      7
#> 143 2015-08-05 2015-07-27 2015-07-31  41  24     10  5.0  9.0   3      8
#> 144 2015-07-29 2015-07-20 2015-07-24  42  23     10  4.0 10.0   3      8
#> 145 2015-07-22 2015-07-13 2015-07-17  41  24     10  5.0 10.0   4      6
#> 146 2015-07-14 2015-07-06 2015-07-10  42  23     11  4.0 10.0   4      6
#> 147 2015-07-08 2015-06-29 2015-07-03  42  23     11  5.0 10.0   4      5
#> 148 2015-07-01 2015-06-22 2015-06-26  42  23     10  5.0 10.0   4      6
#> 149 2015-06-24 2015-06-15 2015-06-19  41  23     10  5.0 10.0   4      7
#> 150 2015-06-17 2015-06-08 2015-06-12  41  23     11  5.0  9.0   5      6
#> 151 2015-06-10 2015-06-01 2015-06-05  41  24     10  6.0  9.0   4      6
#> 152 2015-06-02 2015-05-26 2015-05-29  40  23     10  6.0 10.0   5      6
#> 153 2015-05-27 2015-05-18 2015-05-22  39  23     11  7.0  9.0   4      7
#> 154 2015-05-20 2015-05-11 2015-05-15  40  22     11  6.0  9.0   5      7
#> 155 2015-05-12 2015-05-04 2015-05-08  41  23     12  4.0  9.0   5      6
#> 156 2015-05-06 2015-04-27 2015-04-30  42  24     10  4.0  9.0   4      7
#> 157 2015-04-29 2015-04-20 2015-04-24  42  24     10  4.0  9.0   4      7
#> 158 2015-04-22 2015-04-13 2015-04-17  42  24     10  4.0  8.0   6      6
#> 159 2015-04-15 2015-04-07 2015-04-10  42  24     10  5.0  8.0   5      6
#> 160 2015-04-08 2015-03-30 2015-04-02  42  23     10  4.0  9.0   6      6
#> 161 2015-03-31 2015-03-23 2015-03-27  42  23     10  4.0  9.0   6      6
#> 162 2015-03-25 2015-03-16 2015-03-20  41  24     10  5.0  8.0   5      7
#> 163 2015-03-18 2015-03-09 2015-03-13  41  24      9  5.0  9.0   6      6
#> 164 2015-03-11 2015-03-02 2015-03-06  41  24      9  5.0  9.0   6      6
#> 165 2015-03-04 2015-02-23 2015-02-27  41  24      9  5.0  9.0   7      5
#> 166 2015-02-25 2015-02-16 2015-02-20  42  24      9  5.0  9.0   6      5
#> 167 2015-02-18 2015-02-09 2015-02-13  43  23     10  5.0  8.0   5      6
#> 168 2015-02-11 2015-02-02 2015-02-06  42  23     10  5.0  9.0   6      5
#> 169 2015-02-04 2015-01-26 2015-01-30  42  23     10  4.0 10.0   5      6
#> 170 2015-01-28 2015-01-19 2015-01-23  42  23     10  3.0  9.0   6      7
#> 171 2015-01-21 2015-01-12 2015-01-16  42  22     10  3.0 10.0   6      7
#> 172 2015-01-14 2015-01-05 2015-01-09  42  22     10  4.0 10.0   6      6
#> 173 2015-01-07 2015-01-02 2015-01-05  42  22     11  3.0 10.0   5      7
#> 174 2015-01-01 2015-12-15 2015-12-19  42  23     10  2.0 11.0   5      7
#> 175 2014-12-16 2014-12-08 2014-12-12  43  23     10  2.0 10.0   5      7
#> 176 2014-12-10 2014-12-01 2014-12-05  41  24     11  2.0  9.0   6      7
#> 177 2014-12-03 2014-11-24 2014-11-28  41  24     11  2.0  9.0   6      7
#> 178 2014-11-26 2014-11-17 2014-11-21  42  23     10  2.0 10.0   6      7
#> 179 2014-11-19 2014-11-10 2014-11-14  42  22     10  3.0  9.0   7      7
#> 180 2014-11-12 2014-11-03 2014-11-07  41  23      9  3.0 10.0   8      6
#> 181 2014-11-05 2014-10-27 2014-10-31  41  22     10  3.0  9.0   8      7
#> 182 2014-10-29 2014-10-20 2014-10-24  41  23     10  2.0  9.0   8      7
#> 183 2014-10-22 2014-10-13 2014-10-17  41  24      9  2.0  9.0   8      7
#> 184 2014-10-15 2014-10-06 2014-10-10  40  23     10  3.0  9.0   8      7
#> 185 2014-10-07 2014-09-29 2014-10-02  41  23     10  2.0  9.0   8      7
#> 186 2014-10-01 2014-09-22 2014-09-26  42  23      9  2.0  8.0   9      7
#> 187 2014-09-23 2014-09-15 2014-09-19  42  22      8  2.0  9.0  10      7
#> 188 2014-09-17 2014-09-08 2014-09-12  41  24      9  3.0  9.0   7      7
#> 189 2014-09-10 2014-09-01 2014-09-05  42  23      9  3.0  9.0   7      7
#> 190 2014-09-03 2014-08-25 2014-08-29  41  24     10  3.0 10.0   5      7
#> 191 2014-08-27 2014-08-18 2014-08-22  42  24     10  3.0  9.0   5      7
#> 192 2014-08-20 2014-08-11 2014-08-15  42  24     10  3.0  9.0   5      7
#> 193 2014-08-13 2014-08-04 2014-08-08  42  24     10  4.0  9.0   5      6
#> 194 2014-08-06 2014-07-28 2014-08-01  43  24      9  4.0  9.0   5      6
#> 195 2014-07-30 2014-07-21 2014-07-25  43  22     10  3.0 10.0   5      7
#> 196 2014-07-23 2014-07-14 2014-07-18  43  22     10  3.0 10.0   5      7
#> 197 2014-07-16 2014-07-07 2014-07-11  41  23     11  3.0  9.0   6      7
#> 198 2014-07-09 2014-06-30 2014-07-04  41  23     11  3.0  9.0   6      7
#> 199 2014-07-02 2014-06-23 2014-06-27  41  23     10  4.0  9.0   7      6
#> 200 2014-06-25 2014-06-16 2014-06-20  41  23     10  4.0  9.0   7      6
#> 201 2014-06-17 2014-06-10 2014-06-13  39  23     10  4.0 10.0   7      7
#> 202 2014-06-11 2014-06-02 2014-06-06  39  23      9  4.0 10.0   8      7
#> 203 2014-06-04 2014-05-26 2014-05-30  38  23     10  4.0 10.0   8      7
#> 204 2014-05-27 2014-05-19 2014-05-23  40  25     10  4.0 10.0   6      5
#> 205 2014-05-21 2014-05-12 2014-05-16  40  24     10  4.0 10.0   6      6
#> 206 2014-05-14 2014-05-05 2014-05-09  40  24     10  4.0 10.0   6      6
#> 207 2014-05-07 2014-04-28 2014-05-02  41  23     10  4.0 10.0   5      7
#> 208 2014-04-29 2014-04-22 2014-04-25  41  24     10  4.0  9.0   6      6
#> 209 2014-04-23 2014-04-14 2014-04-17  41  24     10  4.0 10.0   5      6
#> 210 2014-04-15 2014-04-07 2014-04-11  41  23     10  4.0 10.0   6      6
#> 211 2014-04-09 2014-03-31 2014-04-04  42  23     10  4.0 10.0   5      6
#> 212 2014-04-02 2014-03-24 2014-03-28  42  23      9  4.0 10.0   6      6
#> 213 2014-03-26 2014-03-17 2014-03-21  41  23     10  4.0 10.0   5      7
#> 214 2014-03-19 2014-03-10 2014-03-14  42  23      9  4.0 10.0   5      7
#> 215 2014-03-12 2014-03-03 2014-03-07  42  23      9  4.0 10.0   5      7
#> 216 2014-03-05 2014-02-24 2014-02-28  40  22     10  4.0 11.0   6      7
#> 217 2014-02-26 2014-02-17 2014-02-21  41  22     10  5.0 10.0   5      7
#> 218 2014-02-19 2014-02-10 2014-02-14  40  24     10  4.0 10.0   5      7
#> 219 2014-02-12 2014-02-03 2014-02-07  41  24     10  4.0 10.0   4      7
#> 220 2014-02-05 2014-01-27 2014-01-31  42  24     10  4.0  9.0   4      7
#> 221 2014-01-29 2014-01-20 2014-01-24  42  23     11  4.0  9.0   4      7
#> 222 2014-01-22 2014-01-13 2014-01-17  41  23     10  5.0 10.0   4      7
#> 223 2014-01-15 2014-01-06 2014-01-10  40  23     10  5.0 10.0   5      7
#> 224 2014-01-08 2014-01-02 2014-01-03  41  23     10  5.0 10.0   5      6
#> 225 2014-01-01 2014-12-16 2014-12-20  42  24     10  5.0 10.0   4      5
#> 226 2013-12-17 2013-12-09 2013-12-13  42  24      9  4.0 10.0   5      6
#> 227 2013-12-11 2013-12-02 2013-12-06  43  24      9  3.0 10.0   5      6
#> 228 2013-12-04 2013-11-25 2013-11-29  42  23     10  3.0 10.0   5      7
#> 229 2013-11-20 2013-11-11 2013-11-15  41  24     10  4.0 10.0   5      6
#> 230 2013-11-13 2013-11-04 2013-11-08  41  24      9  4.0 10.0   5      7
#> 231 2013-11-06 2013-10-28 2013-11-01  42  25      9  3.0  9.0   5      7
#> 232 2013-10-30 2013-10-21 2013-10-25  42  24      9  4.0  9.0   5      7
#> 233 2013-10-23 2013-10-14 2013-10-18  43  25      8  4.0  9.0   5      6
#> 234 2013-10-16 2013-10-07 2013-10-11  44  24      8  4.0  8.0   6      6
#> 235 2013-10-09 2013-09-30 2013-10-04  45  24      8  3.0  8.0   6      6
#> 236 2013-10-01 2013-09-23 2013-09-27  42  25      8  4.0  9.0   5      7
#>     respondents
#> 1          2507
#> 2          1502
#> 3          2505
#> 4          2506
#> 5          2501
#> 6          1003
#> 7          2507
#> 8          2505
#> 9          1282
#> 10         2501
#> 11         1504
#> 12         2501
#> 13         2003
#> 14         2504
#> 15         2501
#> 16         1566
#> 17         2500
#> 18         2502
#> 19         2508
#> 20         1789
#> 21         2505
#> 22         2506
#> 23         2005
#> 24         2500
#> 25         2503
#> 26         2501
#> 27         2001
#> 28         2005
#> 29         2006
#> 30         2501
#> 31         2505
#> 32         2001
#> 33         2503
#> 34         2501
#> 35         2501
#> 36         2507
#> 37         2503
#> 38         2501
#> 39         2507
#> 40         2502
#> 41         2501
#> 42         2503
#> 43         1507
#> 44         2502
#> 45         2505
#> 46         2004
#> 47         2501
#> 48         2001
#> 49         2502
#> 50         2506
#> 51         2004
#> 52         2502
#> 53         2002
#> 54         2007
#> 55         2502
#> 56         2504
#> 57         2501
#> 58         2504
#> 59         2505
#> 60         2502
#> 61         2500
#> 62         2502
#> 63         2502
#> 64         2501
#> 65         2502
#> 66         2504
#> 67         2503
#> 68         2501
#> 69         2001
#> 70         2504
#> 71         2502
#> 72         2506
#> 73         2501
#> 74         2503
#> 75         2505
#> 76         2503
#> 77         2503
#> 78         2503
#> 79         2504
#> 80         2505
#> 81         2004
#> 82         2503
#> 83         2506
#> 84         2504
#> 85         2501
#> 86         2501
#> 87         2503
#> 88         2504
#> 89         2501
#> 90         2501
#> 91         2508
#> 92         2508
#> 93         2502
#> 94         2503
#> 95         2504
#> 96         2501
#> 97         2503
#> 98         2505
#> 99         2501
#> 100        2502
#> 101        2006
#> 102        2501
#> 103        2002
#> 104        2501
#> 105        2502
#> 106        2504
#> 107        2506
#> 108        2004
#> 109        2005
#> 110        2501
#> 111        2505
#> 112        2501
#> 113        2503
#> 114        2502
#> 115        2501
#> 116        2500
#> 117        2501
#> 118        2505
#> 119        2501
#> 120        2503
#> 121        1503
#> 122        1501
#> 123        2504
#> 124        2502
#> 125        2506
#> 126        2501
#> 127        2502
#> 128        2504
#> 129        2503
#> 130        2504
#> 131        2501
#> 132        2502
#> 133        2502
#> 134        2505
#> 135        2506
#> 136        2501
#> 137        2501
#> 138        2503
#> 139        2504
#> 140        2500
#> 141        2500
#> 142        2500
#> 143        2501
#> 144        2501
#> 145        2501
#> 146        2503
#> 147        2502
#> 148        2505
#> 149        2501
#> 150        2502
#> 151        2504
#> 152        2502
#> 153        2503
#> 154        2003
#> 155        2502
#> 156        2001
#> 157        2506
#> 158        2502
#> 159        2004
#> 160        2001
#> 161        2505
#> 162        2501
#> 163        2500
#> 164        2504
#> 165        2502
#> 166        2501
#> 167        2502
#> 168        2503
#> 169        2510
#> 170        2015
#> 171        2502
#> 172        2504
#> 173        1002
#> 174        2504
#> 175        2504
#> 176        2503
#> 177        2502
#> 178        2501
#> 179        2501
#> 180        2502
#> 181        2504
#> 182        2504
#> 183        2504
#> 184        2505
#> 185        2004
#> 186        2501
#> 187        2501
#> 188        2503
#> 189        2502
#> 190        2504
#> 191        2504
#> 192        2502
#> 193        2500
#> 194        2501
#> 195        2501
#> 196        2505
#> 197        2503
#> 198        2014
#> 199        2502
#> 200        2504
#> 201        2001
#> 202        2501
#> 203        2002
#> 204        2503
#> 205        2501
#> 206        2504
#> 207        2001
#> 208        2002
#> 209        2002
#> 210        2501
#> 211        2502
#> 212        2501
#> 213        2502
#> 214        2500
#> 215        2503
#> 216        2500
#> 217        2502
#> 218        2502
#> 219        2502
#> 220        2501
#> 221        2503
#> 222        2502
#> 223        2503
#> 224        1003
#> 225        2508
#> 226        2505
#> 227        2504
#> 228        2507
#> 229        2501
#> 230        2501
#> 231        2500
#> 232        2501
#> 233        2500
#> 234        2501
#> 235        2000
#> 236        2499