Scrape surveys from wahlrecht.de

Scrapes survey tables and performs sanitation to output tidy data

Scrape Bavarian regional polls

Scrape Lower Saxony regional polls

scrape_wahlrecht(address = "https://www.wahlrecht.de/umfragen/emnid.htm",
  parties = c("CDU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW", "AFD",
  "SONSTIGE"))

scrape_by(address = "https://www.wahlrecht.de/umfragen/landtage/bayern.htm",
  parties = c("CSU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW", "AFD",
  "SONSTIGE"))

  scrape_ltw(address = "https://www.wahlrecht.de/umfragen/landtage/niedersachsen.htm",
  parties = c("CDU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW", "AFD",
  "SONSTIGE"), ind_row_remove = -c(1:2))

Arguments

address
http-address from which tables should be scraped.

parties
A character vector containing names of parties to collapse.

ind_row_remove
Negative vector of rows that will be skipped at the begining.

Examples

library(coalitions)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

# select a polling agency from .pollster_df that should be scraped ...
coalitions:::.pollster_df

#> # A tibble: 7 x 2
#>   pollster   address                                             
#>   <chr>      <chr>                                               
#> 1 allensbach https://www.wahlrecht.de/umfragen/allensbach.htm    
#> 2 emnid      https://www.wahlrecht.de/umfragen/emnid.htm         
#> 3 forsa      https://www.wahlrecht.de/umfragen/forsa.htm         
#> 4 fgw        https://www.wahlrecht.de/umfragen/politbarometer.htm
#> 5 gms        https://www.wahlrecht.de/umfragen/gms.htm           
#> 6 infratest  https://www.wahlrecht.de/umfragen/dimap.htm         
#> 7 insa       https://www.wahlrecht.de/umfragen/insa.htm          

# ... here we choose Forsa
address <- coalitions:::.pollster_df %>% filter(pollster == "forsa") %>% pull(address)
scrape_wahlrecht(address = address) %>% slice(1:5)

#> Warning:  1 failed to parse.

#> Warning:  3 failed to parse.

#> # A tibble: 5 x 11
#>   date       start      end          cdu   spd greens   fdp  left   afd others
#>   <date>     <date>     <date>     <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1 2018-08-18 2018-08-13 2018-08-17   30.   18.    15.    9.   10.   14.     4.
#> 2 2018-08-11 2018-08-06 2018-08-10   30.   18.    15.   10.    9.   14.     4.
#> 3 2018-08-04 2018-07-30 2018-08-03   31.   18.    13.   10.    9.   14.     5.
#> 4 2018-07-28 2018-07-23 2018-07-27   32.   18.    13.    9.    9.   14.     5.
#> 5 2018-07-21 2018-07-16 2018-07-20   31.   18.    13.    8.   10.   15.     5.
#> # ... with 1 more variable: respondents <dbl>

# Niedersachsen
scrape_ltw() %>% slice(1:5)

#> # A tibble: 5 x 12
#>   pollster      date       start      end          cdu   spd greens   fdp  left
#>   <chr>         <date>     <date>     <date>     <dbl> <dbl>  <dbl> <dbl> <dbl>
#> 1 forsa         2018-02-25 2018-02-25 2018-02-25   33.  33.0  10.0     8.  6.00
#> 2 fgw           2017-10-12 2017-10-12 2017-10-12   33.  34.5   9.00    9.  5.00
#> 3 insa          2017-10-09 2017-10-09 2017-10-09   32.  33.0  10.0    10.  5.00
#> 4 fgw           2017-10-05 2017-10-05 2017-10-05   33.  33.0   9.00   10.  5.00
#> 5 infratestdim… 2017-10-05 2017-10-05 2017-10-05   34.  34.0   8.50    8.  4.50
#> # ... with 3 more variables: afd <dbl>, others <dbl>, respondents <dbl>

# Hessen
scrape_ltw("http://www.wahlrecht.de/umfragen/landtage/hessen.htm", ind_row_remove=-c(1)) %>%
 slice(1:5)

#> # A tibble: 5 x 12
#>   pollster      date       start      end          cdu   spd greens   fdp  left
#>   <chr>         <date>     <date>     <date>     <dbl> <dbl>  <dbl> <dbl> <dbl>
#> 1 infratestdim… 2018-06-21 2018-06-21 2018-06-21   31.   22.    14.    7.    7.
#> 2 forschungsgr… 2018-06-12 2018-06-12 2018-06-12   31.   25.    13.    8.    8.
#> 3 insa          2018-05-17 2018-05-17 2018-05-17   33.   24.    13.    7.    8.
#> 4 forschungsgr… 2018-03-22 2018-03-22 2018-03-22   31.   26.    13.    7.    8.
#> 5 forsa         2018-02-25 2018-02-25 2018-02-25   33.   23.    14.    8.    7.
#> # ... with 3 more variables: afd <dbl>, others <dbl>, respondents <dbl>