Scrape surveys from wahlrecht.de

Scrapes survey tables and performs sanitation to output tidy data

scrape_wahlrecht(address = "https://www.wahlrecht.de/umfragen/emnid.htm",
  parties = c("CDU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW",
  "AFD", "SONSTIGE"))

  scrape_by(address = "https://www.wahlrecht.de/umfragen/landtage/bayern.htm",
  parties = c("CSU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW",
  "AFD", "SONSTIGE"))

  scrape_ltw(address = "https://www.wahlrecht.de/umfragen/landtage/niedersachsen.htm",
  parties = c("CDU", "SPD", "GRUENE", "FDP", "LINKE", "PIRATEN", "FW",
  "AFD", "SONSTIGE"), ind_row_remove = -c(1:2))

Arguments

address
http-address from which tables should be scraped.

parties
A character vector containing names of parties to collapse.

ind_row_remove
Negative vector of rows that will be skipped at the beginning.

Examples

library(coalitions)
library(dplyr)
# select a polling agency from .pollster_df that should be scraped ...
coalitions:::.pollster_df

#> # A tibble: 7 x 2
#>   pollster   address                                             
#>   <chr>      <chr>                                               
#> 1 allensbach https://www.wahlrecht.de/umfragen/allensbach.htm    
#> 2 emnid      https://www.wahlrecht.de/umfragen/emnid.htm         
#> 3 forsa      https://www.wahlrecht.de/umfragen/forsa.htm         
#> 4 fgw        https://www.wahlrecht.de/umfragen/politbarometer.htm
#> 5 gms        https://www.wahlrecht.de/umfragen/gms.htm           
#> 6 infratest  https://www.wahlrecht.de/umfragen/dimap.htm         
#> 7 insa       https://www.wahlrecht.de/umfragen/insa.htm          

# ... here we choose Forsa
address <- coalitions:::.pollster_df %>% filter(pollster == "forsa") %>% pull(address)
scrape_wahlrecht(address = address) %>% slice(1:5)

#> Warning:  1 failed to parse.

#> Warning:  3 failed to parse.

#>         date      start        end cdu spd greens fdp left afd others
#> 1 2019-03-04 2019-02-25 2019-03-01  30  16     20   9    8  11      6
#> 2 2019-02-23 2019-02-18 2019-02-22  30  16     20   9    7  12      6
#> 3 2019-02-16 2019-02-11 2019-02-15  29  17     21  10    6  12      5
#> 4 2019-02-09 2019-02-04 2019-02-08  31  15     19  10    8  12      5
#> 5 2019-02-02 2019-01-28 2019-02-01  32  15     19   9    8  12      5
#>   respondents
#> 1        2502
#> 2        2501
#> 3        2501
#> 4        2500
#> 5        2501

# Niedersachsen
scrape_ltw() %>% slice(1:5)

#>         pollster       date      start        end cdu  spd greens fdp left afd
#> 1          forsa 2019-02-11 2019-02-11 2019-02-11  30 28.0     17   7    5   8
#> 2 infratestdimap 2018-11-22 2018-11-22 2018-11-22  28 26.0     24   6    4   9
#> 3           insa 2018-11-21 2018-11-21 2018-11-21  27 26.0     17  10    5  12
#> 4          forsa 2018-02-25 2018-02-25 2018-02-25  33 33.0     10   8    6   6
#> 5            fgw 2017-10-12 2017-10-12 2017-10-12  33 34.5      9   9    5   7
#>   others respondents
#> 1    5.0        1010
#> 2    3.0        1006
#> 3    3.0        1053
#> 4    4.0        1004
#> 5    2.5        1001

# Hessen
scrape_ltw("http://www.wahlrecht.de/umfragen/landtage/hessen.htm", ind_row_remove=-c(1)) %>%
 slice(1:5)

#>                 pollster       date      start        end cdu spd greens fdp
#> 1 forschungsgruppewahlen 2018-10-25 2018-10-25 2018-10-25  28  20     20   8
#> 2                   insa 2018-10-23 2018-10-23 2018-10-23  26  21     21   7
#> 3         infratestdimap 2018-10-18 2018-10-18 2018-10-18  26  21     20   9
#> 4 forschungsgruppewahlen 2018-10-18 2018-10-18 2018-10-18  26  20     22   8
#> 5 forschungsgruppewahlen 2018-10-03 2018-10-03 2018-10-03  29  23     18   6
#>   left afd others respondents
#> 1    8  12      4        1025
#> 2    8  13      4        1004
#> 3    8  12      4        1002
#> 4    8  12      4        1035
#> 5    8  13      3        1038